if-->wait race condition
Putting in a CONDITIONAL wait to combat a race condition in the distributed file system.  It first checks to see if the file exists. If it doesn't, it waits .1 seconds, if it still doesn't, it waits 1.0 seconds.  Then it raises high holy hell about the file not existing.  

So if it does indeed exist, we've only incurred the cost of running a single existence check and an if, so it will only do the wait if it already HAS hit the race condition.