prelim elk docker-compose
Believed-basically-working docker-compose.yml.  This brings up ELK in
a configuration that the security genreally is turned on but the ssl
checking is turned off.  This gets us JUST far enough that the
commands that set the main root password seem to work.

This commit is to put a stake in the ground.  This was based on a
cdrohook config file that was a bit older, so this will need to get
merged in with current versions of it.

Signed-off-by: Craig P Steffen <craigsteffen@gmail.com>