USGS SSL keys Dec 12
From David Watkins on slack on December 10.  I copied the public SSL
key that they'll apparently be using, and also copied in a text file
copying his instructions.

We won't need this in our testing, but we'll want to incorporate this
as we roll it out.  I'm putting it in this branch as a separate commit
on the assumption that this will become the running branch since it
contains all the ELK stuff.  If that doesn't turn out, then this can
be cherry-picked easily enough.