ELK+UIUC_CM slightly works
With this version of the docker-compose.yml, all the UIUC containers
come up, and the two ELK containers come up as well.  With this
configuration I was able to log into the elasticsearch container and
at least start the password setup utility.

So I'm not going to say that it's *working*, but as far as elk is
concnered it works at least as far as ELK ever has at NCSA.