UIUC+ELK seems working
The previous version of the config files produced containers that were
superficially working, but kibana was remaining in the "starting"
state, and was not working.

This is a walk back several configuration versions of the
docker-compose config file and then bringing the kibana and elastic
search into synch.

This version of all of those files brings up all the containers,
including the UIUC containers and the elk containers, which now
indicate that they're "heahlthy".  It's not clear exactly which of the
changes is the critical one.  Also these versions are going back
before the big merge with main from the repo, so they will have to be
re-merged.

But this is the best ELK we have so far.