docker_compose.yml changes from main
This was the almost-working version of the docker-compose.yml merged
between the main branch from a couple of days ago and the not-quite
working ELK configuration.  The differences in this file against the
now checked-in version of the docker-compose.yml from this branch need
to be merged into that file to minimize the ugly merging that will
need to bring this file into line with what's in main.